Treat Content-Format as a number
Treat the Content-Format option as a number - this is an incompatible API
change to libCoap.

Clients will now correctly handle Content-Formats > 255, or Content-Format
options with zero-padding.

NSDL-C seems to have some oddities in its Content-Format handling - it has
some lingering assumptions about "0=text/plain" being the default, which
is no longer true in the final CoRE/CoAP RFCs. This should be looked at;
no functional change is made to this by this commit.

Change-Id: I4c25741964e562178b31508d30184b68783eadc2

Conflicts:
	nsdl-c/sn_nsdl_lib.h
	source/libCoap/src/sn_coap_builder.c