A slightly better fix for sn_coap_protocol, to prevent segfault
Use the mbed-client-config-file instead of defines in makefile.
This makes the libCoap_protocol.cpp to have defines after sn_config.h,
which prevents segfault, which may be missed by test framework.

The valgrind still spots horrors when running this test, so this change
is not yet ready for prime time as is.