protocol: consolidate error paths in sn_coap_handle_blockwise_message()
The input message was freed on most error cases in
sn_coap_handle_blockwise_message(), but not in all of them. This causes
a) memory leaks
b) extreme pain on refactoring, as the unit tests actually require
that the memory leak is there.

Fix this by moving the cleanup after error to the call site and fix
the unit test fall out this causes.

As a side step, fix the sn_coap_protocol_delete_retransmission_by_token()
test which accessed the token memory past given message buffer. Why this
surfaced only now, is likely caused by the test having a memory error
emulation turned on and different amount of allocations popped up the
old bug in test code.