blockwise: reduce heap footprint
If the size1 option present in the first received block message then the allocation can be done before hand.
In case size1 is not present then payload is reallocated and only one entry is stored in the list.
This will reduce the use of heap since there is no need to store every block as it's own entry.

User is now responsible of freeing the data by calling sn_coap_protocol_block_remove() and must not free the payload separately.